"Fixs the CSV importing into table issues."

+++ db/migrate/20161204175652_create_trips.rb
@@ -1,16 +1,16 @@
 class CreateTrips < ActiveRecord::Migration[5.0]
   def change
     create_table :trips do |t|
-      t.integer :duration
+      t.float :duration
       t.string :start_date
       t.string :start_station_name
-      t.integer :start_station_id
+      t.float :start_station_id
       t.string :end_date
       t.string :end_station_name
-      t.integer :end_station_id
-      t.integer :bike_id
+      t.float :end_station_id
+      t.float :bike_id
       t.string :subscription_type
-      t.integer :zip_code
+      t.float :zip_code
 
       t.timestamps null: false
     end
